Modifying the layout design created a separate profile activity ...
