+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".DonorActivity"
-            android:parentActivityName=".MainActivity"></activity>
+            android:name=".ui.DonorActivity"
+            android:parentActivityName=".ui.MainActivity" />
+        <activity android:name=".ui.ProfileActivity"></activity>
     </application>
 
 </manifest>
