--- app/src/main/java/com/android/mvnshrikanth/theblooddonor/DonorFragment.java
@@ -1,29 +1,0 @@
-package com.android.mvnshrikanth.theblooddonor;
-
-import android.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-
-public class DonorFragment extends Fragment {
-    private static final String LOG_TAG = DonorFragment.class.getSimpleName();
-
-    public DonorFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_donor, container, false);
-    }
-}
