+++ app/src/main/java/com/android/mvnshrikanth/theblooddonor/adapters/BloodDataAdapter.java
@@ -1,9 +1,12 @@
-package com.android.mvnshrikanth.theblooddonor;
+package com.android.mvnshrikanth.theblooddonor.adapters;
 
 import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.android.mvnshrikanth.theblooddonor.R;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -18,7 +21,9 @@ public class BloodDataAdapter extends RecyclerView.Adapter<BloodDataAdapter.MyVi
 
     @Override
     public MyViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        return null;
+        View view = LayoutInflater.from(parent.getContext())
+                .inflate(R.layout.item_blood_data, parent, false);
+        return new MyViewHolder(view);
     }
 
     @Override
@@ -26,9 +31,11 @@ public class BloodDataAdapter extends RecyclerView.Adapter<BloodDataAdapter.MyVi
 
     }
 
+
+    //TODO Change the item count in the adapter.
     @Override
     public int getItemCount() {
-        return 0;
+        return 1;
     }
 
     public class MyViewHolder extends RecyclerView.ViewHolder {
