+++ app/src/main/java/com/android/mvnshrikanth/theblooddonor/ui/DonorActivity.java
@@ -1,8 +1,10 @@
-package com.android.mvnshrikanth.theblooddonor;
+package com.android.mvnshrikanth.theblooddonor.ui;
 
 import android.app.FragmentTransaction;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import com.android.mvnshrikanth.theblooddonor.R;
 
 public class DonorActivity extends AppCompatActivity {
     private static final String LOG_TAG = DonorActivity.class.getSimpleName();
