+++ app/src/main/java/com/android/mvnshrikanth/theblooddonor/ui/MainActivity.java
@@ -1,0 +1,18 @@
+package com.android.mvnshrikanth.theblooddonor.ui;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import com.android.mvnshrikanth.theblooddonor.R;
+
+public class MainActivity extends AppCompatActivity {
+    private static final String LOG_TAG = MainActivity.class.getSimpleName();
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+    }
+}
