+++ app/src/main/java/com/android/mvnshrikanth/theblooddonor/ui/ProfileActivity.java
@@ -1,6 +1,5 @@
-package com.android.mvnshrikanth.theblooddonor;
+package com.android.mvnshrikanth.theblooddonor.ui;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.CardView;
@@ -11,11 +10,13 @@ import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.Spinner;
 
+import com.android.mvnshrikanth.theblooddonor.R;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class MainActivity extends AppCompatActivity {
-    private static final String LOG_TAG = MainActivity.class.getSimpleName();
+public class ProfileActivity extends AppCompatActivity {
+
     @BindView(R.id.spinner_sex)
     Spinner spinner_sex;
     @BindView(R.id.spinner_blood_type)
@@ -28,16 +29,12 @@ public class MainActivity extends AppCompatActivity {
     LinearLayout linearLayout_scrollView_container;
     @BindView(R.id.button_save)
     Button button_save;
-    @BindView(R.id.button_my_donations)
-    Button button_my_donations;
-    @BindView(R.id.button_search_donors)
-    Button button_search_donors;
 
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_profile);
         ButterKnife.bind(this);
 
         ArrayAdapter<CharSequence> adapterSex =
@@ -65,14 +62,6 @@ public class MainActivity extends AppCompatActivity {
                 TransitionManager.beginDelayedTransition(linearLayout_scrollView_container);
                 cardView_editable.setVisibility(View.GONE);
                 cardView_view_only.setVisibility(View.VISIBLE);
-            }
-        });
-
-        button_my_donations.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, DonorActivity.class);
-                startActivity(intent);
             }
         });
