+++ app/src/main/res/layout/activity_donor.xml
@@ -4,4 +4,4 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.android.mvnshrikanth.theblooddonor.DonorActivity" />
+    tools:context="com.android.mvnshrikanth.theblooddonor.ui.DonorActivity" />
