+++ app/src/main/res/layout/fragment_donor.xml
@@ -2,16 +2,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.android.mvnshrikanth.theblooddonor.DonorFragment">
+    tools:context="com.android.mvnshrikanth.theblooddonor.ui.DonorFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView_blood_data"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/app_name" />
 
 </FrameLayout>
